Conflicts:
	includes/class.project.php

+++ js/callbacks/deletesearches.php
@@ -11,8 +11,9 @@ $baseurl = dirname(dirname($baseurl)) .'/' ;
 if (Cookie::has('flyspray_userid') && Cookie::has('flyspray_passhash')) {
     $user = new User(Cookie::val('flyspray_userid'));
     $user->check_account_ok();
-    
-    $db->Query('DELETE FROM {searches} WHERE id = ? AND user_id = ?', array(Get::num('id'), $user->id));
+    if (!$user->isAnon()) {
+        $db->Query('DELETE FROM {searches} WHERE id = ? AND user_id = ?', array(Get::num('id'), $user->id));
+    }
 }
 
 ?>
